+++ data/api/mirakl.json
@@ -260,6 +260,62 @@
         }
       }
     },
+    "GetInvoices":{
+      "httpMethod": "GET",
+      "uri": "/api/invoices",
+      "summary": "List accounting documents",
+      "parameters": {
+        "shop": {
+          "required": false,
+          "type": "integer",
+          "description": "[optional] Identifier of the shop for filtering",
+          "location": "query",
+          "sentAs": "shop"
+        },
+        "startDate": {
+          "required": false,  
+          "type": "object",
+          "instanceOf" : "DateTime",
+          "description": "creation date for filtering. Format: 'yyyy-MM-dd'T'HH:mm:ss'",
+          "location": "query",
+          "sentAs": "start_date",
+          "filters" : [
+            {
+              "method": "date_format",
+              "args" : ["@value", "Y-m-d\\TH:i:s\\Z"]
+            }
+          ]
+        },
+        "endDate": {
+          "required": false,  
+          "type": "object",
+          "instanceOf" : "DateTime",
+          "description": "creation date for filtering. Format: 'yyyy-MM-dd'T'HH:mm:ss'",
+          "location": "query",
+          "sentAs": "end_date",
+          "filters" : [
+            {
+              "method": "date_format",
+              "args" : ["@value", "Y-m-d\\TH:i:s\\Z"]
+            }
+          ]
+        },
+        "type": {
+          "required": false,
+          "type": "string",
+          "description": "[optional] [ALL, AUTO_INVOICE, MANUAL_INVOICE, MANUAL_CREDIT] Type of the accounting document for filtering. Default value is AUTO_INVOICE which represent invoices generated by shop payment.",
+          "location": "query",
+          "sentAs": "type"
+        },
+        "currency": {
+          "required": false,
+          "type": "string",
+          "description": "[optional] Currency iso code for filtering",
+          "location": "query",
+          "sentAs": "currency"
+        }
+      }
+    },
     "DocumentTypesDto": {
       "httpMethod": "GET",
       "uri": "/api/documents",
